(#3) Make identifiers great again
Because of the broken tokenizer, when I added the test files, I had to change
the names of some variables anb functions so that they aren't interpreted as
keywords.
Since the tokenizer has been fix recently, I changed these names to what they
were supposed to be.

I also removed the respective warning from the README